Revert sanitize middleware
